+++ setup.py
@@ -2,8 +2,9 @@ from setuptools import setup, find_packages
 
 setup(
     name="streamlit_view",
-    version="1.0.0",
-    packages=find_packages(),
+    version="1.0.2",
+    packages=find_packages(where="src"),
+    package_dir={"": "src"},
     install_requires=[
         "streamlit",
         "openai",
